refactoring from class to functional base
